pyproject.toml
[tool.poetry]
name = "image-ai-utils-server"
version = "0.0.3"
description = ""
authors = ["Sergey Morozov <sergey@morozov.top>", "neonfuzz"]

[tool.poetry.dependencies]
python = ">=3.9,<3.10" # 3.9 required for gfpgan which requires older version of numpy
diffusers = "^0.6.0"
transformers = "^4.21.1"
scipy = "^1.9.0"
fastapi = "^0.79.1"
gunicorn = "^20.1.0"
uvicorn = "^0.18.2"
torch = { version="^1.12.1", source="torch" }
pydantic = "^1.9.2"
Pillow = "^9.2.0"
gfpgan = "^1.3.4"
realesrgan = "^0.2.5.0"
numpy = "1.20.3"
python-dotenv = "^0.20.0"
colorlog = "^6.7.0"
websockets = "^10.3"
httpx = "^0.23.0"
aiofiles = "^22.1.0"

[[tool.poetry.source]]
name = "torch"
url = "https://download.pytorch.org/whl/cu116"
secondary = true


[tool.poetry.group.dev.dependencies]
snakeviz = "^2.1.1"
pyinstaller = "^5.4.1"
requests = "^2.28.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
